add debug to makefiles

+++ build/moduletop.mk
@@ -10,32 +10,36 @@ headers:
 		echo "";\
 		echo "------------------- WORKING DIR ===> $$dirx"; \
 		echo ""; \
-		make -C $$dirx headers || exit 1; \
+		$(MAKE) -C $$dirx headers || exit 1; \
 	done;
 
 exe: libs
 	@for dirx in $(DIRS); do \
-		echo ""; \
-		echo "------------------ WORKING DIR ===> $$dirx"; \
-		echo "";\
-		make -C $$dirx exe || exit 1;\
+	echo ""; \
+	echo "------------------ WORKING DIR ===> $$dirx"; \
+	echo "";\
+	$(MAKE) -C $$dirx exe || exit 1;\
 	done;
 
 libs: $(MODLIB)
 	@for dirx in $(DIRS); do \
-		echo ""; \
-		echo "------------------ WORKING DIR ===> $$dirx"; \
-		echo ""; \
-		make -C $$dirx libs || exit 1; \
-	done
+	echo ""; \
+	echo "------------------ WORKING DIR ===> $$dirx"; \
+	echo "";\
+	$(MAKE) -C $$dirx libs || exit 1;\
+	done;
 
 $(MODLIB): $(OBJECT_C:.o=.lo) $(OBJECT_CXX:.o=.lo)
+ifneq ($(MODLIB),)
 	@echo ""
 	@echo "----- Building library : " $@
-	$(LIBTOOL) --tag=CC --mode=link $(CC) $(CFLAGS) -o lib$@.la $? -rpath $(LIBDIR) $(LD_LIBS) $(LD_LIBS_PATH) $(LIBS) $(LIBS_PATH)
-	echo ""
+	$(LIBTOOL) --tag=CC --mode=link $(CC) $(CFLAGS) -o lib$@.la $? \
+	-rpath $(INSTALL_LIBDIR) $(LD_LIBS) $(LD_LIBS_PATH) $(LIBS) $(LIBS_PATH)
+	@echo ""
+endif
 
 install : $(MODLIB)
+ifneq ($(MODLIB),)
 	@echo ""
 	@echo "-------------- Installing " lib$(MODLIB)  " to " $(INSTALL_LIBDIR)
 	@if [ ! -d $(INSTALL_LIBDIR) ]; then \
@@ -43,27 +47,30 @@ install : $(MODLIB)
 		mkdir -p $(INSTALL_LIBDIR) || exit -1;\
 	fi
 	$(LIBTOOL) --tag=CC --mode=install install -c lib$(MODLIB).la $(INSTALL_LIBDIR)/lib$(MODLIB).la
-#$(LIBTOOL) --finish .libs
+endif
 	@echo ""
 	@echo "----------------- INSTALLING ----------------"
 	@for dirx in $(DIRS); do \
 		echo "";\
 		echo "------------------ WORKING DIR ===> $$dirx"; \
 		echo "";\
-		make -C $$dirx install || exit 1;\
+		$(MAKE) -C $$dirx install || exit 1;\
 	done;
 	@echo "----------------- DONE -----------------------";
 	@echo ""
+# $(LIBTOOL) --finish $(INSTALL_LIBDIR)
 
 uninstall :
+ifneq ($(MODLIB),)
 	@echo ""
 	@echo "---------------- Uninstall -----------------"
 	rm -rf $(INSTALL_LIBDIR)/lib$(MODLIB).*
+endif
 	@for dirx in $(DIRS); do \
 		echo ""; \
 		echo "------------------ WORKING DIR ===> $$dirx"; \
 		echo ""; \
-		make -C $$dirx uninstall; \
+		$(MAKE) -C $$dirx uninstall; \
 	done;
 	@echo "------------------ DONE ---------------------"
 	@echo ""
@@ -71,16 +78,21 @@ uninstall :
 clean :
 	@echo ""
 	@echo "---------------- CLEANING -----------------"
-	rm -rf $(CLEAN_TARGETS) $(INSTALL_LIBDIR)/lib$(MODLIB).*
+	@rm -rf $(CLEAN_TARGETS) $(INSTALL_LIBDIR)/lib$(MODLIB).*
 	@for dirx in $(DIRS); do \
 		echo ""; \
 		echo "------------------ WORKING DIR ===> $$dirx"; \
 		echo ""; \
-		make -C $$dirx clean; \
+		$(MAKE) -C $$dirx clean; \
 	done;
 	@echo "----------------- DONE -----------------------";
 	@echo ""
-
+debug:
+	@$(MAKE) clean
+	@$(MAKE) -e BUILD_ENV=debug
+	@for dirx in $(DIRS); do \
+		$(MAKE) -C $$dirx debug; \
+	done;
 # <====== COMPLING RULES ========>
 $(OBJECT_C:.o=.lo) : $(OBJECT_C)
 $(OBJECT_C) : %.o: %.c
@@ -88,4 +100,4 @@ $(OBJECT_C) : %.o: %.c
 
 $(OBJECT_CXX:.o=.lo) : $(OBJECT_CXX)
 $(OBJECT_CXX) : %.o: %.cpp
-	$(LIBTOOL) --tag=CC --mode=compile $(CXX) $(CFLAGS) $(INC_PATH) -c $<
+	$(LIBTOOL) --tag=CXX --mode=compile $(CXX) $(CFLAGS) $(INC_PATH) -c $<
